Stub for field dependency manager
